pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reconx"
version = "1.0.0"
description = "AI-powered automated recon orchestrator for bug bounty and pentesting"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Yeswanth"}
]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "dnspython>=2.4",
    "rich>=13.0",
    "httpx>=0.27",
    "openai>=1.0",
    "ollama>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
reconx = "reconx.cli:main"

[tool.setuptools.packages.find]
include = ["reconx*"]
